added a module external_api.py and moved the func get_actual_currency and rewrote the tests for it
