+++ src/utils.py
@@ -1,9 +1,7 @@
 import json
 from typing import Any
 
-import requests
-
-from src.decorators import retry
+from src.external_api import get_actual_currency
 from src.my_logger import logger
 
 
@@ -40,32 +38,6 @@ def get_amount_transaction(transaction: dict) -> Any:
         message = "Транзакция в указанной валюте не обрабатывается"
         logger.warning(message)
         return ValueError(message)
-    except Exception as ex:
-        error = f"{ex.__class__.__name__}: {ex}"
-        logger.error(error)
-        return error
-
-
-@retry(3)
-def get_actual_currency(currency: str) -> Any:
-    """
-    Получает актуальный курс валюты из внешнего источника.
-
-    :param currency: Код валюты (например, "RUB").
-    :return: Актуальный курс валюты.
-    """
-    try:
-        url = "https://www.cbr-xml-daily.ru/daily_json.js"
-        response = requests.get(url)
-        if response.status_code == 200:
-            currency_info = response.json()
-            return currency_info["Valute"][currency]["Value"]
-        return []
-
-    except KeyError:
-        logger.warning('Функция работает только с "RUB", "USD" и "EUR"')
-        return 'KeyError: Функция работает только с "RUB", "USD" и "EUR"'
-
     except Exception as ex:
         error = f"{ex.__class__.__name__}: {ex}"
         logger.error(error)
